Renamed model directory to DTO
