+++ EverTea/User/BackEnd/Advanced Weather/AdvancedWeatherForecastApp/src/main/java/com/evertea/AdvancedWeatherForecastApp/DTO/WeatherData.java
@@ -1,4 +1,4 @@
-package com.evertea.AdvancedWeatherForecastApp.Model;
+package com.evertea.AdvancedWeatherForecastApp.DTO;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
