+++ EverTea/User/BackEnd/Advanced Weather/AdvancedWeatherForecastApp/src/main/java/com/evertea/AdvancedWeatherForecastApp/controller/WeatherController.java
@@ -1,14 +1,12 @@
 package com.evertea.AdvancedWeatherForecastApp.controller;
 
-import com.evertea.AdvancedWeatherForecastApp.Model.WeatherData;
+import com.evertea.AdvancedWeatherForecastApp.DTO.WeatherData;
 import com.evertea.AdvancedWeatherForecastApp.service.WeatherService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/weather")
