+++ EverTea/User/BackEnd/Advanced Weather/AdvancedWeatherForecastApp/src/main/java/com/evertea/AdvancedWeatherForecastApp/repo/NotificationRepository.java
@@ -1,6 +1,6 @@
 package com.evertea.AdvancedWeatherForecastApp.repo;
 
-import com.evertea.AdvancedWeatherForecastApp.Model.WeatherNotification;
+import com.evertea.AdvancedWeatherForecastApp.DTO.WeatherNotification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.stereotype.Repository;
